# EvDrivers/API_SWAGGER_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: EV Battery Swap Management System API
  description: |
    Comprehensive API for EV Battery Swap Management System with monthly billing based on distance usage.
    
    ## Authentication
    Use POST `/api/auth/login` to get JWT token, then include in Authorization header:
    ```
    Authorization: Bearer jwt-token-xxxxx
    ```
    
    ## Monthly Billing System
    - Battery swaps do NOT charge money immediately (cost = 0)
    - System tracks distance usage automatically via odometer readings
    - Monthly billing based on service plans and distance overage
    - Payment processed at end of month with usage reset
    
    ## Database Credentials (for testing)
    - **Admin**: admin@evswap.com / admin123
    - **Driver**: minh.driver@gmail.com / driver123  
    - **Staff**: duc.staff@evswap.com / staff123
    
  version: 1.0.0
  contact:
    name: EV Battery Swap Team
    email: dev@evswap.com
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/payments/current-bill-status/user/{userId}:
    get:
      tags: [Payment & Billing]
      summary: Get current bill status for all user contracts
      description: Overview of all active contracts with current billing status
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: driver001
      responses:
        '200':
          description: Bill status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BillStatus'
                  totalActiveContracts:
                    type: integer
                  totalPendingAmount:
                    type: number
...
  /api/contracts/user/{userId}:
    get:
      tags: [Contract Management]
      summary: Get user contracts
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            example: driver001
      responses:
        '200':
          description: User contracts retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Contract'
...
  /api/users/{id}:
    get:
      tags: [User Management]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: driver001
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: admin001
        email:
          type: string
          example: admin@evswap.com
        name:
          type: string
          example: Duc Anh Nguyen
        role:
          type: string
          enum: [Admin, Staff, EV Driver]
          example: Admin
        phone:
          type: string
          example: '0901234567'
        status:
          type: string
          example: active
...
    BillStatus:
      type: object
      properties:
        contractId:
          type: integer
        contractNumber:
          type: string
        currentMonth:
          type: string
        planName:
          type: string
        planType:
          type: string
        baseDistance:
          type: integer
        basePrice:
          type: number
        isUnlimited:
          type: boolean
        distanceUsed:
          type: number
        baseFee:
          type: number
        overageDistance:
          type: number
        overageFee:
          type: number
        currentTotalFee:
          type: number
        usageProgress:
          type: number
        warningLevel:
          type: string
          enum: [NONE, LOW, MEDIUM, HIGH]
        warningMessage:
          type: string
        daysRemainingInMonth:
          type: integer
...
    Contract:
      type: object
      properties:
        contractId:
          type: integer
        vehicleId:
          type: integer
        contractNumber:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        planId:
          type: integer
        planType:
          type: string
        planName:
          type: string
        monthlyFee:
          type: number
        baseDistance:
          type: integer
        depositFee:
          type: number
        description:
          type: string
        isUnlimited:
          type: boolean
        usedSwaps:
          type: integer
        usedDistance:
          type: number
        monthlyBaseFee:
          type: number
        monthlyOverageFee:
          type: number
        monthlyTotalFee:
          type: number
        currentMonth:
          type: string
...